+++ components/Tooltip/vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cypress-design/vue-tooltip
 
+## 0.0.2
+
+### Patch Changes
+
+- [`e3eb9c7`](https://github.com/cypress-io/cypress-design/commit/e3eb9c7fee2d7a6e0a773e85ed4b73be04d83587) Thanks [@elevatebart](https://github.com/elevatebart)! - downgrade typescript to 4.4 for compat with app
+
 ## 0.0.1
 
 ### Patch Changes